Move DummyWrapper daemon to an own python module
We are goint to use the DummyWrapper in other tests too. Therefore move
it to a common extra module.
