+++ tests/test_scan_and_result.py
@@ -31,8 +31,9 @@ import defusedxml.lxml as secET
 
 from defusedxml.common import EntitiesForbidden
 
-from ospd.ospd import OSPDaemon
 from ospd.errors import OspdCommandError
+
+from .helper import DummyWrapper
 
 
 class Result(object):
@@ -48,162 +49,6 @@ class Result(object):
         self.qod = ''
         for name, value in kwargs.items():
             setattr(self, name, value)
-
-
-class DummyWrapper(OSPDaemon):
-    def __init__(self, results, checkresult=True):
-        super().__init__()
-        self.checkresult = checkresult
-        self.results = results
-
-    def check(self):
-        return self.checkresult
-
-    @staticmethod
-    def get_custom_vt_as_xml_str(vt_id, custom):
-        return '<custom><mytest>static test</mytest></custom>'
-
-    @staticmethod
-    def get_params_vt_as_xml_str(vt_id, vt_params):
-        return (
-            '<params><param id="abc" type="string">'
-            '<name>ABC</name><description>Test ABC</description>'
-            '<default>yes</default></param>'
-            '<param id="def" type="string">'
-            '<name>DEF</name><description>Test DEF</description>'
-            '<default>no</default></param></params>'
-        )
-
-    @staticmethod
-    def get_refs_vt_as_xml_str(vt_id, vt_refs):
-        response = (
-            '<refs><ref type="cve" id="CVE-2010-4480"/>'
-            '<ref type="url" id="http://example.com"/></refs>'
-        )
-        return response
-
-    @staticmethod
-    def get_dependencies_vt_as_xml_str(vt_id, vt_dependencies):
-        response = (
-            '<dependencies>'
-            '<dependency vt_id="1.3.6.1.4.1.25623.1.0.50282" />'
-            '<dependency vt_id="1.3.6.1.4.1.25623.1.0.50283" />'
-            '</dependencies>'
-        )
-
-        return response
-
-    @staticmethod
-    def get_severities_vt_as_xml_str(vt_id, severities):
-        response = (
-            '<severities><severity cvss_base="5.0" cvss_'
-            'type="cvss_base_v2">AV:N/AC:L/Au:N/C:N/I:N/'
-            'A:P</severity></severities>'
-        )
-
-        return response
-
-    @staticmethod
-    def get_detection_vt_as_xml_str(
-        vt_id, detection=None, qod_type=None, qod=None
-    ):
-        response = '<detection qod_type="package">some detection</detection>'
-
-        return response
-
-    @staticmethod
-    def get_summary_vt_as_xml_str(vt_id, summary):
-        response = '<summary>Some summary</summary>'
-
-        return response
-
-    @staticmethod
-    def get_affected_vt_as_xml_str(vt_id, affected):
-        response = '<affected>Some affected</affected>'
-
-        return response
-
-    @staticmethod
-    def get_impact_vt_as_xml_str(vt_id, impact):
-        response = '<impact>Some impact</impact>'
-
-        return response
-
-    @staticmethod
-    def get_insight_vt_as_xml_str(vt_id, insight):
-        response = '<insight>Some insight</insight>'
-
-        return response
-
-    @staticmethod
-    def get_solution_vt_as_xml_str(
-        vt_id, solution, solution_type=None, solution_method=None
-    ):
-        response = '<solution>Some solution</solution>'
-
-        return response
-
-    @staticmethod
-    def get_creation_time_vt_as_xml_str(
-        vt_id, creation_time
-    ):  # pylint: disable=arguments-differ
-        response = '<creation_time>%s</creation_time>' % creation_time
-
-        return response
-
-    @staticmethod
-    def get_modification_time_vt_as_xml_str(
-        vt_id, modification_time
-    ):  # pylint: disable=arguments-differ
-        response = (
-            '<modification_time>%s</modification_time>' % modification_time
-        )
-
-        return response
-
-    def exec_scan(self, scan_id, target):
-        time.sleep(0.01)
-        for res in self.results:
-            if res.result_type == 'log':
-                self.add_scan_log(
-                    scan_id,
-                    res.host or target,
-                    res.hostname,
-                    res.name,
-                    res.value,
-                    res.port,
-                )
-            if res.result_type == 'error':
-                self.add_scan_error(
-                    scan_id,
-                    res.host or target,
-                    res.hostname,
-                    res.name,
-                    res.value,
-                    res.port,
-                )
-            elif res.result_type == 'host-detail':
-                self.add_scan_host_detail(
-                    scan_id,
-                    res.host or target,
-                    res.hostname,
-                    res.name,
-                    res.value,
-                )
-            elif res.result_type == 'alarm':
-                self.add_scan_alarm(
-                    scan_id,
-                    res.host or target,
-                    res.hostname,
-                    res.name,
-                    res.value,
-                    res.port,
-                    res.test_id,
-                    res.severity,
-                    res.qod,
-                )
-            else:
-                raise ValueError(res.result_type)
 
 
 class ScanTestCase(unittest.TestCase):
